+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## 2020-27-10
 
+* Add helper class for packet handling from free5gc v.3.0.4
+* Remove the need for mongodb
+* The PDU Session and the Data plane processing are now disaggregated
+* Maintain the UPD Socket open
+
+## 2020-27-10
+
 * Update readme for more clarification on the installation of the gNB and the usage
 * Add new parameters to the configuration file (security, plnm, snssai)